Almost all of my updates and typo catches
So I worked on this version and when I got to Sally in the Alley, there's no old hooker spitting blood. It seems like more work for me to make changes and typos in a document and then have you transfer those changes into the final epub version. I'd rather be able to make changes to the final document and have you accept or decline from that document. I only say this because typos and errors are still in part one after we proof-read it a couple of times. Anyhow, I'm sure we'll go over it tomorrow. Peace out.